internal: rm catalog, pg, postgres packages
Remove internal/postgresql/ast
Remove internal/postgresql/rewrite
Remove internal/postgresql/validate
